chore: reorganize repo for multiple packages
