+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "schemas"]
-	path = schemas
+	path = providers/openfeature-provider-flagd/schemas
 	url = https://github.com/open-feature/schemas
