+++ hooks/README.md
@@ -1,0 +1,4 @@
+# OpenFeature Python Hooks
+Hooks are a mechanism whereby application developers can add arbitrary behavior to flag evaluation.
+They operate similarly to middleware in many web frameworks. Please see the
+[spec](https://openfeature.dev/specification/sections/hooks) for more details.
