+++ hooks/openfeature-hooks-opentelemetry/openfeature/contrib/hooks/opentelemetry/__init__.py
@@ -1,9 +1,8 @@
 import json
 
-from opentelemetry import trace
-
 from openfeature.flag_evaluation import FlagEvaluationDetails
 from openfeature.hook import Hook, HookContext
+from opentelemetry import trace
 
 OTEL_EVENT_NAME = "feature_flag"
