+++ providers/openfeature-provider-flagd/openfeature/contrib/provider/flagd/defaults.py
