+++ providers/openfeature-provider-flagd/openfeature/contrib/provider/flagd/flag_type.py
