+++ providers/openfeature-provider-flagd/tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 
 from openfeature import api
-from openfeature.contrib.providers.flagd import FlagdProvider
+from openfeature.contrib.provider.flagd import FlagdProvider
 
 
 @pytest.fixture()
