--- pyproject.toml
@@ -1,79 +1,0 @@
-# pyproject.toml
-[build-system]
-requires      = ["setuptools>=61.0.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "openfeature_sdk_contrib"
-version = "0.1.0"
-description = "Contributions around the Python OpenFeature SDK, such as providers and hooks"
-readme = "readme.md"
-authors = [{ name = "OpenFeature", email = "openfeature-core@groups.io" }]
-license = { file = "LICENSE" }
-classifiers = [
-    "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
-]
-keywords = []
-dependencies = []
-requires-python = ">=3.8"
-
-[project.optional-dependencies]
-dev = ["pip-tools", "pytest", "pre-commit"]
-
-[project.urls]
-Homepage = "https://github.com/open-feature/python-sdk-contrib"
-
-[tool.ruff]
-exclude = [
-    ".git",
-    ".venv",
-    "__pycache__",
-    "venv",
-    "openfeature/contrib/providers/flagd/proto/*"
-]
-target-version = "py38"
-
-[tool.ruff.lint]
-select = [
-    "A",
-    "B",
-    "C4",
-    "C90",
-    "E",
-    "F",
-    "FLY",
-    "FURB",
-    "I",
-    "LOG",
-    "N",
-    "PERF",
-    "PGH",
-    "PLC",
-    "PLR0913",
-    "PLR0915",
-    "RUF",
-    "S",
-    "SIM",
-    "T10",
-    "T20",
-    "UP",
-    "W",
-    "YTT",
-]
-ignore = [
-    "E501", # the formatter will handle any too long line
-]
-preview = true
-
-[tool.ruff.lint.per-file-ignores]
-"tests/**/*" = ["S101"]
-
-[tool.ruff.lint.pylint]
-max-args = 6
-max-statements = 30
-
-[tool.ruff.lint.pyupgrade]
-# Preserve types, even if a file imports `from __future__ import annotations`.
-keep-runtime-typing = true
