+++ ruff.toml
@@ -1,0 +1,54 @@
+exclude = [
+    ".git",
+    ".venv",
+    "__pycache__",
+    "venv",
+    "openfeature/contrib/providers/flagd/proto/*"
+]
+target-version = "py38"
+
+[lint]
+select = [
+    "A",
+    "B",
+    "C4",
+    "C90",
+    "E",
+    "F",
+    "FLY",
+    "FURB",
+    "I",
+    "LOG",
+    "N",
+    "PERF",
+    "PGH",
+    "PLC",
+    "PLR0913",
+    "PLR0915",
+    "RUF",
+    "S",
+    "SIM",
+    "T10",
+    "T20",
+    "UP",
+    "W",
+    "YTT",
+]
+ignore = [
+    "E501", # the formatter will handle any too long line
+]
+preview = true
+
+[lint.isort]
+known-first-party = ["openfeature"]
+
+[lint.per-file-ignores]
+"**/tests/**/*" = ["S101"]
+
+[lint.pylint]
+max-args = 6
+max-statements = 30
+
+[lint.pyupgrade]
+# Preserve types, even if a file imports `from __future__ import annotations`.
+keep-runtime-typing = true
